--- ESGF.Sudoku.Spark.RecursiveSearch/ESGF.Sudoku.Spark.RecursiveSearch.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.808.10
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ESGF.Sudoku.Spark.RecursiveSearch", "ESGF.Sudoku.Spark.RecursiveSearch.csproj", "{5EE28271-7F91-43FA-9673-881B8824676F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5EE28271-7F91-43FA-9673-881B8824676F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5EE28271-7F91-43FA-9673-881B8824676F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5EE28271-7F91-43FA-9673-881B8824676F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5EE28271-7F91-43FA-9673-881B8824676F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {89E735A0-2B7A-4A25-A50D-31B69C71496F}
-	EndGlobalSection
-EndGlobal
