--- ESGF.Sudoku.Spark.RecursiveSearch/packages.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="DlxLib" version="1.3.0.0" targetFramework="net452" />
-  <package id="System.Collections" version="4.0.0" targetFramework="net452" />
-  <package id="System.Collections.Immutable" version="1.1.37" targetFramework="net452" />
-  <package id="System.Diagnostics.Debug" version="4.0.0" targetFramework="net452" />
-  <package id="System.Globalization" version="4.0.0" targetFramework="net452" />
-  <package id="System.Linq" version="4.0.0" targetFramework="net452" />
-  <package id="System.Resources.ResourceManager" version="4.0.0" targetFramework="net452" />
-  <package id="System.Runtime" version="4.0.0" targetFramework="net452" />
-  <package id="System.Runtime.Extensions" version="4.0.0" targetFramework="net452" />
-  <package id="System.Threading" version="4.0.0" targetFramework="net452" />
-</packages>
