+++ ESGF.Sudoku.Spark.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31019.35
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ESGF.Sudoku.Spark.RecursiveSearch", "ESGF.Sudoku.Spark.RecursiveSearch\ESGF.Sudoku.Spark.RecursiveSearch.csproj", "{7BD022F1-5C24-406A-A8E8-77D9E7B4D750}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Spark.CSharp.Examples", "Microsoft.Spark.CSharp.Examples\Microsoft.Spark.CSharp.Examples.csproj", "{FEAC21C3-F576-4254-A6F6-60481571E062}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F547DDA1-7F03-4E10-8919-9DFF1E40E58F}"
@@ -12,20 +10,22 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ESGF.Sudoku.Spark.Dancinlinks", "ESGF.Sudoku.Spark.Dancinlinks\ESGF.Sudoku.Spark.Dancinlinks.csproj", "{2DB8F662-5B84-43CF-88D1-752A48F7F8DA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7BD022F1-5C24-406A-A8E8-77D9E7B4D750}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7BD022F1-5C24-406A-A8E8-77D9E7B4D750}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7BD022F1-5C24-406A-A8E8-77D9E7B4D750}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7BD022F1-5C24-406A-A8E8-77D9E7B4D750}.Release|Any CPU.Build.0 = Release|Any CPU
 		{FEAC21C3-F576-4254-A6F6-60481571E062}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{FEAC21C3-F576-4254-A6F6-60481571E062}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FEAC21C3-F576-4254-A6F6-60481571E062}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FEAC21C3-F576-4254-A6F6-60481571E062}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2DB8F662-5B84-43CF-88D1-752A48F7F8DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2DB8F662-5B84-43CF-88D1-752A48F7F8DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2DB8F662-5B84-43CF-88D1-752A48F7F8DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2DB8F662-5B84-43CF-88D1-752A48F7F8DA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
